requirements.txt
asgiref==3.8.1
Django==5.2.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
psycopg2==2.9.10
sqlparse==0.5.3
requests
Pillow>=10.0.0
